redefined the "chargebacks/revoked" and "denied".
change the column frauds into chargebacks for the other site, but call it
revoked for site seeme.com, and make the frauds actually the denied for site
seeme.com. and we only cut the sales according to chargebacks/revoked but
not from denied. For an example, say there are 3 sales and 1 chargebacks/revoked
and 2 denied, then the nets should be (3-1)=2, but not (3-2)=1 nor (3-2-1)=0.

Signed-off-by: Ray Allen <ray@ray-laptop.(none)>